requirements.txt
pytube
requests
python-dotenv
langchain_groq
crewai
pyttsx3
youtube-transcript-api
python-dotenv

